incremental work: adding some comments set control variable's to int64 for some reason ATB seems to mess up the variable if the step size is rather big (3600).

+++ Capture/eye.py
@@ -272,12 +272,12 @@ def eye(src,size,g_pool):
             bar.camera_state[control["name"]]=c_bool(control["value"])
             bar.add_var("Camera/"+control["name"],bar.camera_ctl[control["name"]])
         elif control["type"]=="(int)":
-            bar.camera_ctl[control["name"]]=c_int(control["value"])
-            bar.camera_state[control["name"]]=c_int(control["value"])
+            bar.camera_ctl[control["name"]]=c_int64(control["value"])
+            bar.camera_state[control["name"]]=c_int64(control["value"])
             bar.add_var("Camera/"+control["name"],bar.camera_ctl[control["name"]],max=control["max"],min=control["min"],step=control["step"])
         elif control["type"]=="(menu)":
-            bar.camera_ctl[control["name"]]=c_int(control["value"])
-            bar.camera_state[control["name"]]=c_int(control["value"])
+            bar.camera_ctl[control["name"]]=c_int64(control["value"])
+            bar.camera_state[control["name"]]=c_int64(control["value"])
             bar.add_var("Camera/"+control["name"],bar.camera_ctl[control["name"]],max=control["max"],min=control["min"],step=1)
         else:
             pass
